--- app/src/main/java/jp/co/archive_asia/onedaycouplediary/SplashFragment.kt
@@ -1,35 +1,0 @@
-package jp.co.archive_asia.onedaycouplediary
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import androidx.navigation.fragment.findNavController
-import jp.co.archive_asia.onedaycouplediary.databinding.FragmentSplashBinding
-
-class SplashFragment : Fragment () {
-
-    private var _binding: FragmentSplashBinding? = null
-    private val binding get() = _binding!!
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        _binding = FragmentSplashBinding.inflate(inflater, container, false)
-        return binding.root
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-
-        findNavController().navigate(R.id.action_splashFragment_to_introFragment)
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
-}
