+++ app/src/main/java/jp/co/archive_asia/onedaycouplediary/view/auth/IntroFragment.kt
@@ -1,30 +1,14 @@
 package jp.co.archive_asia.onedaycouplediary.view.auth
 
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
 import jp.co.archive_asia.onedaycouplediary.R
 import jp.co.archive_asia.onedaycouplediary.databinding.FragmentIntroBinding
+import jp.co.archive_asia.onedaycouplediary.view.BaseFragment
 
-class IntroFragment : Fragment() {
+class IntroFragment : BaseFragment<FragmentIntroBinding>(R.layout.fragment_intro) {
 
-    private var _binding: FragmentIntroBinding? = null
-    private val binding get() = _binding!!
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        _binding = FragmentIntroBinding.inflate(inflater, container, false)
-        return binding.root
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
+    override fun initView() {
+        super.initView()
 
         binding.loginBtn.setOnClickListener {
             findNavController().navigate(R.id.action_introFragment_to_loginFragment)
@@ -35,8 +19,4 @@ class IntroFragment : Fragment() {
         }
     }
 
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
 }
