+++ app/src/main/java/jp/co/archive_asia/onedaycouplediary/view/auth/JoinFragment.kt
@@ -1,28 +1,7 @@
 package jp.co.archive_asia.onedaycouplediary.view.auth
 
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
+import jp.co.archive_asia.onedaycouplediary.R
 import jp.co.archive_asia.onedaycouplediary.databinding.FragmentJoinBinding
+import jp.co.archive_asia.onedaycouplediary.view.BaseFragment
 
-class JoinFragment : Fragment() {
-
-    private var _binding: FragmentJoinBinding? = null
-    private val binding get() = _binding!!
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        _binding = FragmentJoinBinding.inflate(inflater, container, false)
-        return binding.root
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
-}
+class JoinFragment : BaseFragment<FragmentJoinBinding>(R.layout.fragment_join)
