+++ app/src/main/java/jp/co/archive_asia/onedaycouplediary/view/auth/LoginFragment.kt
@@ -1,30 +1,17 @@
 package jp.co.archive_asia.onedaycouplediary.view.auth
 
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
+import androidx.navigation.fragment.findNavController
 import jp.co.archive_asia.onedaycouplediary.R
-import jp.co.archive_asia.onedaycouplediary.databinding.FragmentJoinBinding
 import jp.co.archive_asia.onedaycouplediary.databinding.FragmentLoginBinding
+import jp.co.archive_asia.onedaycouplediary.view.BaseFragment
 
-class LoginFragment : Fragment() {
+class LoginFragment : BaseFragment<FragmentLoginBinding>(R.layout.fragment_login) {
 
-    private var _binding: FragmentLoginBinding? = null
-    private val binding get() = _binding!!
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        _binding = FragmentLoginBinding.inflate(inflater, container, false)
-        return binding.root
+    override fun initView() {
+        super.initView()
+        binding.btnLogin.setOnClickListener {
+            findNavController().navigate(R.id.action_loginFragment_to_calendarFragment)
+        }
     }
 
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
 }
