+++ app/src/main/java/jp/co/archive_asia/onedaycouplediary/view/fragment/CalendarFragment.kt
@@ -1,27 +1,17 @@
 package jp.co.archive_asia.onedaycouplediary.view.fragment
 
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.databinding.DataBindingUtil
 import jp.co.archive_asia.onedaycouplediary.R
 import jp.co.archive_asia.onedaycouplediary.databinding.FragmentCalendarBinding
+import jp.co.archive_asia.onedaycouplediary.view.BaseFragment
 import java.time.LocalDate
 import java.time.format.DateTimeFormatter
 
-class CalendarFragment : Fragment() {
+class CalendarFragment : BaseFragment<FragmentCalendarBinding>(R.layout.fragment_calendar) {
 
-    private lateinit var binding: FragmentCalendarBinding
     lateinit var selectedDate: LocalDate
 
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-
-        binding = DataBindingUtil.inflate(inflater, R.layout.fragment_calendar, container, false)
+    override fun initView() {
+        super.initView()
 
         selectedDate = LocalDate.now()
 
@@ -36,8 +26,6 @@ class CalendarFragment : Fragment() {
             selectedDate = selectedDate.plusMonths(1)
             setMonthView()
         }
-
-        return binding.root
     }
 
     private fun setMonthView() {
