+++ app/src/main/res/layout/fragment_login.xml
@@ -9,6 +9,7 @@
         tools:context=".view.auth.LoginFragment">
 
         <Button
+            android:id="@+id/btn_login"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:text="ログイン"
