+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/splashFragment"
-        android:name="jp.co.archive_asia.onedaycouplediary.SplashFragment"
+        android:name="jp.co.archive_asia.onedaycouplediary.view.SplashFragment"
         android:label="SplashFragment"
         tools:layout="@layout/fragment_splash" >
         <action
